requirements.txt
mlflow>=2.15.0
mlflow-skinny>=2.15.0
mlflow-tracing>=2.15.0
pandas>=2.2.3
PyYAML>=6.0.3
scikit-learn>=1.6.0
seaborn>=0.13.2
tensorflow>=2.18.0
tensorflow-datasets>=4.9.9
numpy>=1.24
Pillow>=12.1.0
streamlit>=1.28.0